requirements.txt
alabaster==0.7.12
argon2-cffi==20.1.0
asn1crypto==0.24.0
astroid==2.4.2
async-generator==1.10
atomicwrites==1.3.0
attrs==20.2.0
autopep8==1.5.4
autoreject==0.2.1
Babel==2.7.0
backcall==0.2.0
bcrypt==3.1.7
bitstring==3.1.6
bleach==3.2.1
blist==1.3.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.21
deap==1.3.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.15.2
ecdsa==0.13.2
entrypoints==0.3
future==0.17.1
idna==2.10
imagesize==1.1.0
importlib-metadata==0.22
ipaddress==1.0.22
ipykernel==5.3.4
ipympl==0.5.8
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.3
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
liac-arff==2.4.0
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mne==0.21.0
mne-bids==0.4
mock==3.0.5
more-itertools==7.2.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
netaddr==0.7.19
netifaces==0.10.9
nose==1.3.7
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.7.1
pathlib2==2.3.4
patsy==0.5.1
paycheck==1.0.2
pbr==5.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psutil==5.6.3
ptitprince==0.2.5
ptyprocess==0.6.0
py==1.8.0
py-expression-eval==0.3.9
pyasn1==0.4.7
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.1
PyHamcrest==2.0.2
pylint==2.6.0
pymanopt==0.2.5
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyriemann==0.2.6
pyrsistent==0.17.3
pytest==5.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scandir==1.10.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
setuptools-scm==3.3.3
six==1.15.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
statsmodels==0.12.0
tabulate==0.8.3
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.5
ujson==1.35
urllib3==1.25.10
virtualenv==16.7.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==0.6.0
